Add the --host option to the cli (#594)
The flag in question allows users to set the host on which the server listens to
a different value from the default 127.0.0.1 interface. This is crucial to run
the sqslite server in a Docker container and expose it to an external network,
for instance, in a sidecar running on a Kubernetes pod in a CI/CD cloud native pipeline.